+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
   their functions to the `HasId` and `MaybeHasId` traits
   - Renamed `native_id`/`try_native_id` to `js_raw_id`/`try_js_raw_id` for consistency with the
     other trait functions
+- Remove `Resolvable` trait, moving its functionality to `MaybeHasId`
 
 0.19.0 (2023-12-20)
 ===================